FastAPI/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
databases[aiosqlite]==0.8.0
sqlalchemy==1.4.53
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0

